# src/main/java/eu/clarin/cmdi/virtualcollectionregistry/rest/BaseResource.java
/*
 * Copyright (C) 2014 CLARIN
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
package eu.clarin.cmdi.virtualcollectionregistry.rest;

import java.io.IOException;
import java.io.InputStream;
import java.io.OutputStream;
import javax.ws.rs.GET;
import javax.ws.rs.Path;
import javax.ws.rs.Produces;
import javax.ws.rs.WebApplicationException;
import javax.ws.rs.container.ResourceContext;
import javax.ws.rs.core.*;
import org.apache.wicket.util.io.IOUtils;
/**
 *
 * @author twagoo
 */
//@Component
@Path("")
public class BaseResource {
    @Context
    private ResourceContext resourceContext;

    /**
     * Serves a short description HTML page at the service root
     *
     * @return
     */
    @GET
    @Produces({MediaType.TEXT_XML})
    public Response getDescription() {
        final StreamingOutput writer = new StreamingOutput() {
            @Override
            public void write(OutputStream output) throws IOException,
                    WebApplicationException {
                try (InputStream is = getClass().getResourceAsStream("/restIndex.html")) {
                    IOUtils.copy(is, output);
                } finally {
                    output.close();
                }
            }
        };
        return Response.ok(writer).type(MediaType.TEXT_HTML).build();
    }

    @Path("/v1/collections")
    public VirtualCollectionsResource getCollectionsV1() {
        final VirtualCollectionsResource resource =
                resourceContext.getResource(VirtualCollectionsResource.class);
        return resource;
    }

    @Path("/v2/collections")
    public VirtualCollectionsResource getCollectionsV2() {
        final VirtualCollectionsResource resource =
                resourceContext.getResource(VirtualCollectionsResource.class);
        return resource;
    }
}